refactor tooltip to reusable jsx component applied to factory methods
