+++ src/components/chloropleth/tooltips/region/hospitalAdmissionsTooltip.tsx
@@ -1,8 +1,8 @@
 import { NextRouter } from 'next/router';
 import { ReactNode } from 'react';
-import styles from '~/components/chloropleth/tooltips/tooltip.module.scss';
 import { SafetyRegionProperties } from '~/components/chloropleth/shared';
 import { createSelectRegionHandler } from '~/components/chloropleth/selectHandlers/createSelectRegionHandler';
+import { Tooltip } from '~/components/chloropleth/tooltips/tooltipContent';
 
 export const createRegionHospitalAdmissionsTooltip = (router: NextRouter) => (
   context: SafetyRegionProperties & { value: number }
@@ -16,20 +16,9 @@ export const createRegionHospitalAdmissionsTooltip = (router: NextRouter) => (
 
   return (
     context && (
-      <div className={styles.escalationTooltip} onClick={onSelectRegion}>
-        <div className={styles.escalationTooltipHeader}>
-          <h3>{context?.vrname}</h3>
-        </div>
-        {
-          <div className={styles.positiveTestedPeopleInfo}>
-            <div className={styles.escalationText}>
-              <strong>
-                {context.value !== undefined ? context.value : '-'}
-              </strong>
-            </div>
-          </div>
-        }
-      </div>
+      <Tooltip title={context.vrname} onSelectRegion={onSelectRegion}>
+        {context.value !== undefined ? context.value : '-'}
+      </Tooltip>
     )
   );
 };
